RPC-468 refactoring code to make it simpler to make changes to individual calculations
