+++ src/main.rs
@@ -1,24 +1,15 @@
 use std::{env, net::UdpSocket, sync::Arc};
 
-use crate::rpc_server::AtlasPriorityFeeEstimatorRpcServer;
 use cadence::{BufferedUdpMetricSink, QueuingMetricSink, StatsdClient};
 use cadence_macros::set_global_default;
 use figment::{providers::Env, Figment};
-use grpc_geyser::GrpcGeyserImpl;
 use jsonrpsee::server::ServerBuilder;
 use jsonrpsee::server::middleware::http::ProxyGetRequestLayer;
-use priority_fee::PriorityFeeTracker;
-use rpc_server::AtlasPriorityFeeEstimator;
 use serde::Deserialize;
 use tracing::{error, info};
-
-mod errors;
-mod grpc_consumer;
-mod grpc_geyser;
-mod priority_fee;
-mod rpc_server;
-mod slot_cache;
-mod solana;
+use atlas_priority_fee_estimator::grpc_geyser::GrpcGeyserImpl;
+use atlas_priority_fee_estimator::priority_fee::PriorityFeeTracker;
+use atlas_priority_fee_estimator::rpc_server::{AtlasPriorityFeeEstimator, AtlasPriorityFeeEstimatorRpcServer};
 
 #[derive(Debug, Deserialize, Clone)]
 struct EstimatorEnv {
