+++ src/priority_fee.rs
@@ -1,126 +1,32 @@
+use crate::errors::TransactionValidationError;
 use crate::grpc_consumer::GrpcConsumer;
+use crate::model::{Fees, MicroLamportPriorityFeeEstimates, PriorityFeesBySlot, SlotPriorityFees};
+use crate::priority_fee_calculation::Calculations;
+use crate::priority_fee_calculation::Calculations::Calculation1;
 use crate::rpc_server::get_recommended_fee;
 use crate::slot_cache::SlotCache;
 use cadence_macros::statsd_count;
 use cadence_macros::statsd_gauge;
 use dashmap::DashMap;
-use serde::Deserialize;
-use serde::Serialize;
 use solana::storage::confirmed_block::Message;
 use solana_program_runtime::compute_budget::ComputeBudget;
 use solana_program_runtime::prioritization_fee::PrioritizationFeeDetails;
 use solana_sdk::instruction::CompiledInstruction;
 use solana_sdk::transaction::TransactionError;
 use solana_sdk::{pubkey::Pubkey, slot_history::Slot};
-use std::collections::HashMap;
 use std::sync::Arc;
-use std::time::{Duration, Instant};
-use tokio::sync::mpsc::{channel, Receiver, Sender};
-use tracing::{debug, error};
+use std::time::Duration;
+use tracing::error;
 use yellowstone_grpc_proto::geyser::subscribe_update::UpdateOneof;
 use yellowstone_grpc_proto::geyser::{SubscribeUpdate, SubscribeUpdateTransactionInfo};
 use yellowstone_grpc_proto::prelude::{MessageHeader, Transaction, TransactionStatusMeta};
 use yellowstone_grpc_proto::solana;
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
-pub enum PriorityLevel {
-    Min,       // 0th percentile
-    Low,       // 25th percentile
-    Medium,    // 50th percentile
-    High,      // 75th percentile
-    VeryHigh,  // 95th percentile
-    UnsafeMax, // 100th percentile
-    Default,   // 50th percentile
-}
-
-impl From<String> for PriorityLevel {
-    fn from(s: String) -> Self {
-        match s.as_str() {
-            "NONE" => PriorityLevel::Min,
-            "LOW" => PriorityLevel::Low,
-            "MEDIUM" => PriorityLevel::Medium,
-            "HIGH" => PriorityLevel::High,
-            "VERY_HIGH" => PriorityLevel::VeryHigh,
-            "UNSAFE_MAX" => PriorityLevel::UnsafeMax,
-            _ => PriorityLevel::Default,
-        }
-    }
-}
-
-impl Into<Percentile> for PriorityLevel {
-    fn into(self) -> Percentile {
-        match self {
-            PriorityLevel::Min => 0,
-            PriorityLevel::Low => 25,
-            PriorityLevel::Medium => 50,
-            PriorityLevel::High => 75,
-            PriorityLevel::VeryHigh => 95,
-            PriorityLevel::UnsafeMax => 100,
-            PriorityLevel::Default => 50,
-        }
-    }
-}
-
-type Percentile = usize;
-
-#[derive(Deserialize, Serialize, Debug, Clone, Default)]
-#[serde(rename_all(serialize = "camelCase", deserialize = "camelCase"))]
-pub struct MicroLamportPriorityFeeEstimates {
-    pub min: f64,
-    pub low: f64,
-    pub medium: f64,
-    pub high: f64,
-    pub very_high: f64,
-    pub unsafe_max: f64,
-}
-
-#[derive(Debug, Clone)]
-struct SlotPriorityFees {
-    fees: Fees,
-    account_fees: DashMap<Pubkey, Fees>,
-}
-type PriorityFeesBySlot = DashMap<Slot, SlotPriorityFees>;
-
-impl SlotPriorityFees {
-    fn new(accounts: Vec<Pubkey>, priority_fee: u64, is_vote: bool) -> Self {
-        let account_fees = DashMap::new();
-        let fees = Fees::new(priority_fee as f64, is_vote);
-        for account in accounts {
-            account_fees.insert(account, fees.clone());
-        }
-        if is_vote {
-            Self { fees, account_fees }
-        } else {
-            Self { fees, account_fees }
-        }
-    }
-}
 
 #[derive(Debug, Clone)]
 pub struct PriorityFeeTracker {
     priority_fees: Arc<PriorityFeesBySlot>,
     compute_budget: ComputeBudget,
     slot_cache: SlotCache,
-    sampling_sender: Sender<(Vec<Pubkey>, bool, bool, Option<u32>)>,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-enum TransactionValidationError {
-    TransactionFailed,
-    TransactionMissing,
-    MessageMissing,
-    InvalidAccount,
-}
-
-impl Into<&str> for TransactionValidationError {
-    fn into(self) -> &'static str {
-        match self {
-            TransactionValidationError::TransactionFailed => "txn_failed",
-            TransactionValidationError::TransactionMissing => "txn_missing",
-            TransactionValidationError::MessageMissing => "message_missing",
-            TransactionValidationError::InvalidAccount => "invalid_pubkey",
-        }
-    }
 }
 
 fn extract_from_meta(
@@ -159,7 +65,7 @@ fn extract_from_transaction(
         .ok_or(TransactionValidationError::MessageMissing)?;
     let is_vote = transaction.is_vote;
 
-    Result::Ok((message, writable_accounts, is_vote))
+    Ok((message, writable_accounts, is_vote))
 }
 
 fn extract_from_message(
@@ -299,199 +205,66 @@ impl GrpcConsumer for PriorityFeeTracker {
 
 impl PriorityFeeTracker {
     pub fn new(slot_cache_length: usize) -> Self {
-        let (sampling_txn, sampling_rxn) = channel::<(Vec<Pubkey>, bool, bool, Option<u32>)>(100);
-
         let tracker = Self {
             priority_fees: Arc::new(DashMap::new()),
             slot_cache: SlotCache::new(slot_cache_length),
             compute_budget: ComputeBudget::default(),
-            sampling_sender: sampling_txn,
         };
-        tracker.poll_fees(sampling_rxn);
+        tracker.poll_fees();
         tracker
     }
 
-    fn poll_fees(&self, mut sampling_rxn: Receiver<(Vec<Pubkey>, bool, bool, Option<u32>)>) {
-        {
-            let priority_fee_tracker = self.clone();
-            // task to run global fee comparison every 1 second
-            tokio::spawn(async move {
-                loop {
-                    tokio::time::sleep(Duration::from_millis(1_000)).await;
-                    priority_fee_tracker.record_general_fees();
-                }
-            });
-        }
-
-        {
-            let priority_fee_tracker = self.clone();
-            // task to poll the queue and run comparison to see what is the diff between algos
-            tokio::spawn(async move {
-                loop {
-                    match sampling_rxn.recv().await {
-                        Some((accounts, include_vote, include_empty_slots, lookback_period)) => priority_fee_tracker
-                            .record_specific_fees(accounts, include_vote, include_empty_slots, lookback_period),
-                        _ => {}
-                    }
-                }
-            });
-        }
+    fn poll_fees(&self) {
+        let priority_fee_tracker = self.clone();
+        // task to run global fee comparison every 1 second
+        tokio::spawn(async move {
+            loop {
+                tokio::time::sleep(Duration::from_millis(1_000)).await;
+                priority_fee_tracker.record_general_fees();
+            }
+        });
     }
 
     fn record_general_fees(&self) {
-        let global_fees = self.calculation1(&vec![], false, false, &None);
-        statsd_gauge!(
-            "min_priority_fee",
-            global_fees.min as u64,
-            "account" => "none"
-        );
-        statsd_gauge!("low_priority_fee", global_fees.low as u64, "account" => "none");
-        statsd_gauge!(
-            "medium_priority_fee",
-            global_fees.medium as u64,
-            "account" => "none"
-        );
-        statsd_gauge!(
-            "high_priority_fee",
-            global_fees.high as u64,
-            "account" => "none"
-        );
-        statsd_gauge!(
-            "very_high_priority_fee",
-            global_fees.very_high as u64,
-            "account" => "none"
-        );
-        statsd_gauge!(
-            "unsafe_max_priority_fee",
-            global_fees.unsafe_max as u64,
-            "account" => "none"
-        );
-        statsd_gauge!(
-            "recommended_priority_fee",
-            get_recommended_fee(global_fees) as u64,
-            "account" => "none"
-        );
-    }
-
-    fn record_specific_fees(
-        &self,
-        accounts: Vec<Pubkey>,
-        include_vote: bool,
-        include_empty_slots: bool,
-        lookback_period: Option<u32>,
-    ) {
-        let old_fee = self.calculation1(&accounts, include_vote, include_empty_slots, &lookback_period);
-        let new_fee = self.calculation2(&accounts, include_vote, include_empty_slots, &lookback_period);
-        let new_fee_last = self.calculation2(&accounts, include_vote, include_empty_slots, &Some(1));
-
-        statsd_gauge!(
-            "min_priority_fee",
-            old_fee.min,
-            "account" => "spec"
-        );
-        statsd_gauge!(
-            "min_priority_fee_new",
-            new_fee.min,
-            "account" => "spec"
-        );
-        statsd_gauge!(
-            "min_priority_fee_last",
-            new_fee_last.min,
-            "account" => "spec"
-        );
-
-        statsd_gauge!("low_priority_fee",
-            old_fee.low,
-            "account" => "spec"
-        );
-        statsd_gauge!("low_priority_fee_new",
-            new_fee.low,
-            "account" => "spec"
-        );
-        statsd_gauge!("low_priority_fee_last",
-            new_fee_last.low,
-            "account" => "spec"
-        );
-
-        statsd_gauge!(
-            "medium_priority_fee",
-            old_fee.medium,
-            "account" => "spec"
-        );
-        statsd_gauge!(
-            "medium_priority_fee_new",
-            new_fee.medium,
-            "account" => "spec"
-        );
-        statsd_gauge!(
-            "medium_priority_fee_last",
-            new_fee_last.medium,
-            "account" => "spec"
-        );
-
-        statsd_gauge!(
-            "high_priority_fee",
-            old_fee.high,
-            "account" => "spec"
-        );
-        statsd_gauge!(
-            "high_priority_fee_new",
-            new_fee.high,
-            "account" => "spec"
-        );
-        statsd_gauge!(
-            "high_priority_fee_last",
-            new_fee_last.high,
-            "account" => "spec"
-        );
-
-        statsd_gauge!(
-            "very_high_priority_fee",
-            old_fee.very_high,
-            "account" => "spec"
-        );
-        statsd_gauge!(
-            "very_high_priority_fee_new",
-            new_fee.very_high,
-            "account" => "spec"
-        );
-        statsd_gauge!(
-            "very_high_priority_fee_last",
-            new_fee_last.very_high,
-            "account" => "spec"
-        );
-
-        statsd_gauge!(
-            "unsafe_max_priority_fee",
-            old_fee.unsafe_max,
-            "account" => "spec"
-        );
-        statsd_gauge!(
-            "unsafe_max_priority_fee_new",
-            new_fee.unsafe_max,
-            "account" => "spec"
-        );
-        statsd_gauge!(
-            "very_high_priority_fee_last",
-            new_fee_last.unsafe_max,
-            "account" => "spec"
-        );
-
-        statsd_gauge!(
-            "recommended_priority_fee",
-            get_recommended_fee(old_fee),
-            "account" => "spec"
-        );
-        statsd_gauge!(
-            "recommended_priority_fee_new",
-            get_recommended_fee(new_fee),
-            "account" => "spec"
-        );
-        statsd_gauge!(
-            "recommended_priority_fee_last",
-            get_recommended_fee(new_fee_last),
-            "account" => "spec"
-        );
+        let global_fees = self.calculate_priority_fee_details(&Calculation1 {
+            accounts: &vec![],
+            include_vote: false,
+            include_empty_slots: false,
+            lookback_period: &None,
+        });
+        if let Ok(global_fees) = global_fees {
+            statsd_gauge!(
+                "min_priority_fee",
+                global_fees.min as u64,
+                "account" => "none"
+            );
+            statsd_gauge!("low_priority_fee", global_fees.low as u64, "account" => "none");
+            statsd_gauge!(
+                "medium_priority_fee",
+                global_fees.medium as u64,
+                "account" => "none"
+            );
+            statsd_gauge!(
+                "high_priority_fee",
+                global_fees.high as u64,
+                "account" => "none"
+            );
+            statsd_gauge!(
+                "very_high_priority_fee",
+                global_fees.very_high as u64,
+                "account" => "none"
+            );
+            statsd_gauge!(
+                "unsafe_max_priority_fee",
+                global_fees.unsafe_max as u64,
+                "account" => "none"
+            );
+            statsd_gauge!(
+                "recommended_priority_fee",
+                get_recommended_fee(global_fees) as u64,
+                "account" => "none"
+            );
+        }
     }
 
     pub fn push_priority_fee_for_txn(
@@ -505,8 +278,10 @@ impl PriorityFeeTracker {
         // for removal later
         let slot_to_remove = self.slot_cache.push_pop(slot);
         if !self.priority_fees.contains_key(&slot) {
-            self.priority_fees
-                .insert(slot, SlotPriorityFees::new(accounts, priority_fee, is_vote));
+            self.priority_fees.insert(
+                slot,
+                SlotPriorityFees::new(slot, accounts, priority_fee, is_vote),
+            );
         } else {
             // update the slot priority fees
             self.priority_fees.entry(slot).and_modify(|priority_fees| {
@@ -525,297 +300,12 @@ impl PriorityFeeTracker {
         }
     }
 
-    // TODO: DKH - both algos should probably be in some enum (like algo1, algo2) and be passed to
-    // this method instead of sending a bool flag. I'll refactor this in next iteration. already too many changes
-    pub fn get_priority_fee_estimates(
+    pub fn calculate_priority_fee_details(
         &self,
-        accounts: Vec<Pubkey>,
-        include_vote: bool,
-        include_empty_slots: bool,
-        lookback_period: Option<u32>,
-        calculation1: bool,
-    ) -> MicroLamportPriorityFeeEstimates {
-        let start = Instant::now();
-        let micro_lamport_priority_fee_estimates = if calculation1 {
-            self.calculation1(&accounts, include_vote, include_empty_slots, &lookback_period)
-        } else {
-            self.calculation2(&accounts, include_vote, include_empty_slots, &lookback_period)
-        };
-
-        let version = if calculation1 { "v1" } else { "v2" };
-        statsd_gauge!(
-            "get_priority_fee_estimates_time",
-            start.elapsed().as_nanos() as u64,
-            "version" => &version
-        );
-        statsd_count!(
-            "get_priority_fee_calculation_version",
-            1,
-            "version" => &version
-        );
-        if let Err(e) = self.sampling_sender.try_send((
-            accounts.to_owned(),
-            include_vote,
-            include_empty_slots,
-            lookback_period.to_owned(),
-        )) {
-            debug!("Did not add sample for calculation, {:?}", e);
-        }
-
-        micro_lamport_priority_fee_estimates
+        calculation: &Calculations,
+    ) -> anyhow::Result<MicroLamportPriorityFeeEstimates> {
+        calculation.get_priority_fee_estimates(&self.priority_fees)
     }
-
-    /*
-    Algo1: given the list of accounts the algorithm will:
-     1. collect all the transactions fees over n slots
-     2. collect all the transaction fees for all the accounts specified over n slots
-     3. will calculate the percentile distributions for each of two groups
-     4. will choose the highest value from each percentile between two groups
-     */
-    fn calculation1(
-        &self,
-        accounts: &Vec<Pubkey>,
-        include_vote: bool,
-        include_empty_slots: bool,
-        lookback_period: &Option<u32>,
-    ) -> MicroLamportPriorityFeeEstimates {
-        let mut account_fees = vec![];
-        let mut transaction_fees = vec![];
-        for (i, slot_priority_fees) in self.priority_fees.iter().enumerate() {
-            if let Some(lookback_period) = lookback_period {
-                if i >= *lookback_period as usize {
-                    break;
-                }
-            }
-            if include_vote {
-                transaction_fees.extend_from_slice(&slot_priority_fees.fees.vote_fees);
-            }
-            transaction_fees.extend_from_slice(&slot_priority_fees.fees.non_vote_fees);
-            for account in accounts {
-                let account_priority_fees = slot_priority_fees.account_fees.get(account);
-                if let Some(account_priority_fees) = account_priority_fees {
-                    if include_vote {
-                        account_fees.extend_from_slice(&account_priority_fees.vote_fees);
-                    }
-                    account_fees.extend_from_slice(&account_priority_fees.non_vote_fees);
-                }
-            }
-        }
-        if include_empty_slots {
-            let lookback = lookback_period.map(|v| v as usize).unwrap_or(self.priority_fees.len());
-            let account_max_size = account_fees.len().max(lookback);
-            let transaction_max_size = transaction_fees.len().max(lookback);
-            // if there are less data than number of slots - append 0s for up to number of slots so we don't overestimate the values
-            account_fees.resize(account_max_size, 0f64);
-            transaction_fees.resize(transaction_max_size, 0f64);
-        }
-
-        let micro_lamport_priority_fee_estimates = MicroLamportPriorityFeeEstimates {
-            min: max_percentile(&mut account_fees, &mut transaction_fees, 0),
-            low: max_percentile(&mut account_fees, &mut transaction_fees, 25),
-            medium: max_percentile(&mut account_fees, &mut transaction_fees, 50),
-            high: max_percentile(&mut account_fees, &mut transaction_fees, 75),
-            very_high: max_percentile(&mut account_fees, &mut transaction_fees, 95),
-            unsafe_max: max_percentile(&mut account_fees, &mut transaction_fees, 100),
-        };
-
-        micro_lamport_priority_fee_estimates
-    }
-
-    /*
-    Algo2: given the list of accounts the algorithm will:
-     1. collect all the transactions fees over n slots
-     2. for each specified account collect the fees and calculate the percentiles
-     4. choose maximum values for each percentile between all transactions and each account
-     */
-    fn calculation2(
-        &self,
-        accounts: &Vec<Pubkey>,
-        include_vote: bool,
-        include_empty_slots: bool,
-        lookback_period: &Option<u32>,
-    ) -> MicroLamportPriorityFeeEstimates {
-        let mut slots_vec = Vec::with_capacity(self.slot_cache.len());
-        self.slot_cache.copy_slots(&mut slots_vec);
-        slots_vec.sort();
-        slots_vec.reverse();
-
-        let slots_vec = slots_vec;
-
-        let lookback = calculate_lookback_size(&lookback_period, slots_vec.len());
-
-        let mut fees = Vec::with_capacity(slots_vec.len());
-        let mut micro_lamport_priority_fee_estimates = MicroLamportPriorityFeeEstimates::default();
-
-        for slot in &slots_vec[..lookback] {
-            if let Some(slot_priority_fees) = self.priority_fees.get(slot) {
-                if include_vote {
-                    fees.extend_from_slice(&slot_priority_fees.fees.vote_fees);
-                }
-                fees.extend_from_slice(&slot_priority_fees.fees.non_vote_fees);
-            }
-        }
-        micro_lamport_priority_fee_estimates =
-            estimate_max_values(&mut fees, micro_lamport_priority_fee_estimates);
-
-        for account in accounts {
-            fees.clear();
-            let mut zero_slots: usize = 0;
-            for slot in &slots_vec[..lookback] {
-                if let Some(slot_priority_fees) = self.priority_fees.get(slot) {
-                    let account_priority_fees = slot_priority_fees.account_fees.get(account);
-                    if let Some(account_priority_fees) = account_priority_fees {
-                        if include_vote {
-                            fees.extend_from_slice(&account_priority_fees.vote_fees);
-                        }
-                        fees.extend_from_slice(&account_priority_fees.non_vote_fees);
-                    }
-                    else {
-                        zero_slots += 1;
-                    }
-                }
-                else {
-                    zero_slots += 1;
-                }
-            }
-            if include_empty_slots {
-                // for all empty slot we need to add a 0
-                fees.resize(fees.len() + zero_slots, 0f64);
-            }
-            micro_lamport_priority_fee_estimates =
-                estimate_max_values(&mut fees, micro_lamport_priority_fee_estimates);
-        }
-        micro_lamport_priority_fee_estimates
-    }
-}
-
-fn estimate_max_values(
-    mut fees: &mut Vec<f64>,
-    mut estimates: MicroLamportPriorityFeeEstimates,
-) -> MicroLamportPriorityFeeEstimates {
-    let vals: HashMap<Percentile, f64> = percentile(
-        &mut fees,
-        &[
-            PriorityLevel::Min.into(),
-            PriorityLevel::Low.into(),
-            PriorityLevel::Medium.into(),
-            PriorityLevel::High.into(),
-            PriorityLevel::VeryHigh.into(),
-            PriorityLevel::UnsafeMax.into(),
-        ],
-    );
-    estimates.min = vals
-        .get(&PriorityLevel::Min.into())
-        .unwrap_or(&estimates.min)
-        .max(estimates.min);
-    estimates.low = vals
-        .get(&PriorityLevel::Low.into())
-        .unwrap_or(&estimates.low)
-        .max(estimates.low);
-    estimates.medium = vals
-        .get(&PriorityLevel::Medium.into())
-        .unwrap_or(&estimates.medium)
-        .max(estimates.medium);
-    estimates.high = vals
-        .get(&PriorityLevel::High.into())
-        .unwrap_or(&estimates.high)
-        .max(estimates.high);
-    estimates.very_high = vals
-        .get(&PriorityLevel::VeryHigh.into())
-        .unwrap_or(&estimates.very_high)
-        .max(estimates.very_high);
-    estimates.unsafe_max = vals
-        .get(&PriorityLevel::UnsafeMax.into())
-        .unwrap_or(&estimates.unsafe_max)
-        .max(estimates.unsafe_max);
-    estimates
-}
-
-fn max(a: f64, b: f64) -> f64 {
-    if a > b {
-        a
-    } else {
-        b
-    }
-}
-
-fn calculate_lookback_size(pref_num_slots: &Option<u32>, max_available_slots: usize) -> usize {
-    max_available_slots.min(
-        pref_num_slots
-            .map(|v| v as usize)
-            .unwrap_or(max_available_slots),
-    )
-}
-
-#[derive(Debug, Clone)]
-pub struct Fees {
-    non_vote_fees: Vec<f64>,
-    vote_fees: Vec<f64>,
-}
-
-impl Fees {
-    pub fn new(fee: f64, is_vote: bool) -> Self {
-        if is_vote {
-            Self {
-                vote_fees: vec![fee],
-                non_vote_fees: vec![],
-            }
-        } else {
-            Self {
-                vote_fees: vec![],
-                non_vote_fees: vec![fee],
-            }
-        }
-    }
-
-    pub fn add_fee(&mut self, fee: f64, is_vote: bool) {
-        if is_vote {
-            self.vote_fees.push(fee);
-        } else {
-            self.non_vote_fees.push(fee);
-        }
-    }
-}
-
-fn max_percentile(
-    account_fees: &mut Vec<f64>,
-    transaction_fees: &mut Vec<f64>,
-    percentile_value: Percentile,
-) -> f64 {
-    let results1 = percentile(account_fees, &[percentile_value]);
-    let results2 = percentile(transaction_fees, &[percentile_value]);
-    max(
-        *results1.get(&percentile_value).unwrap_or(&0f64),
-        *results2.get(&percentile_value).unwrap_or(&0f64),
-    )
-}
-
-// pulled from here - https://www.calculatorsoup.com/calculators/statistics/percentile-calculator.php
-// couldn't find any good libraries that worked
-fn percentile(values: &mut Vec<f64>, percentiles: &[Percentile]) -> HashMap<Percentile, f64> {
-    if values.is_empty() {
-        return HashMap::with_capacity(0);
-    }
-
-    values.sort_by(|a, b| a.partial_cmp(b).unwrap());
-    let n = values.len() as f64;
-
-    percentiles.into_iter().fold(
-        HashMap::with_capacity(percentiles.len()),
-        |mut data, &percentile| {
-            let r = (percentile as f64 / 100.0) * (n - 1.0) + 1.0;
-
-            let val = if r.fract() == 0.0 {
-                values[r as usize - 1]
-            } else {
-                let ri = r.trunc() as usize - 1;
-                let rf = r.fract();
-                values[ri] + rf * (values[ri + 1] - values[ri])
-            };
-            data.insert(percentile, val);
-            data
-        },
-    )
 }
 
 #[cfg(test)]
@@ -833,6 +323,42 @@ mod tests {
         let noop = NopMetricSink {};
         let client = StatsdClient::builder("", noop).build();
         set_global_default(client)
+    }
+
+    fn calculation1(
+        accounts: &Vec<Pubkey>,
+        include_vote: bool,
+        include_empty_slot: bool,
+        lookback_period: &Option<u32>,
+        tracker: &PriorityFeeTracker,
+    ) -> MicroLamportPriorityFeeEstimates {
+        let calc = Calculations::new_calculation1(
+            accounts,
+            include_vote,
+            include_empty_slot,
+            lookback_period,
+        );
+        tracker
+            .calculate_priority_fee_details(&calc)
+            .expect(format!("estimates for calc1 to be valid with {:?}", calc).as_str())
+    }
+
+    fn calculation2(
+        accounts: &Vec<Pubkey>,
+        include_vote: bool,
+        include_empty_slot: bool,
+        lookback_period: &Option<u32>,
+        tracker: &PriorityFeeTracker,
+    ) -> MicroLamportPriorityFeeEstimates {
+        let calc = Calculations::new_calculation2(
+            accounts,
+            include_vote,
+            include_empty_slot,
+            lookback_period,
+        );
+        tracker
+            .calculate_priority_fee_details(&calc)
+            .expect(format!("estimates for calc2 to be valid with {:?}", calc).as_str())
     }
 
     #[tokio::test]
@@ -857,24 +383,13 @@ mod tests {
         }
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation1(&vec![accounts.get(0).unwrap().clone()], false, false, &None);
-        // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
-        let expected_min_fee = 0.0;
-        let expected_low_fee = 25.0;
-        let expected_medium_fee = 50.0;
-        let expected_high_fee = 75.0;
-        let expected_very_high_fee = 95.0;
-        let expected_max_fee = 100.0;
-        assert_eq!(estimates.min, expected_min_fee);
-        assert_eq!(estimates.low, expected_low_fee);
-        assert_eq!(estimates.medium, expected_medium_fee);
-        assert_eq!(estimates.high, expected_high_fee);
-        assert_eq!(estimates.very_high, expected_very_high_fee);
-        assert_eq!(estimates.unsafe_max, expected_max_fee);
-
-
-        // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation1(&vec![accounts.get(0).unwrap().clone()], false, false, &Some(150));
+        let estimates = calculation1(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            false,
+            &None,
+            &tracker,
+        );
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
@@ -890,7 +405,13 @@ mod tests {
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation1(&vec![accounts.get(0).unwrap().clone()], false, true, &None);
+        let estimates = calculation1(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            false,
+            &Some(150),
+            &tracker,
+        );
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
@@ -905,9 +426,36 @@ mod tests {
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
+        // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
+        let estimates = calculation1(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            true,
+            &None,
+            &tracker,
+        );
+        // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
+        let expected_min_fee = 0.0;
+        let expected_low_fee = 25.0;
+        let expected_medium_fee = 50.0;
+        let expected_high_fee = 75.0;
+        let expected_very_high_fee = 95.0;
+        let expected_max_fee = 100.0;
+        assert_eq!(estimates.min, expected_min_fee);
+        assert_eq!(estimates.low, expected_low_fee);
+        assert_eq!(estimates.medium, expected_medium_fee);
+        assert_eq!(estimates.high, expected_high_fee);
+        assert_eq!(estimates.very_high, expected_very_high_fee);
+        assert_eq!(estimates.unsafe_max, expected_max_fee);
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation1(&vec![accounts.get(0).unwrap().clone()], false, true, &Some(150));
+        let estimates = calculation1(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            true,
+            &Some(150),
+            &tracker,
+        );
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 0.0;
@@ -945,24 +493,13 @@ mod tests {
         }
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation2(&vec![accounts.get(0).unwrap().clone()], false, false, &None);
-        // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
-        let expected_min_fee = 0.0;
-        let expected_low_fee = 25.0;
-        let expected_medium_fee = 50.0;
-        let expected_high_fee = 75.0;
-        let expected_very_high_fee = 95.0;
-        let expected_max_fee = 100.0;
-        assert_eq!(estimates.min, expected_min_fee);
-        assert_eq!(estimates.low, expected_low_fee);
-        assert_eq!(estimates.medium, expected_medium_fee);
-        assert_eq!(estimates.high, expected_high_fee);
-        assert_eq!(estimates.very_high, expected_very_high_fee);
-        assert_eq!(estimates.unsafe_max, expected_max_fee);
-
-
-        // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation2(&vec![accounts.get(0).unwrap().clone()], false, false, &Some(150));
+        let estimates = calculation2(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            false,
+            &None,
+            &tracker,
+        );
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
@@ -978,7 +515,13 @@ mod tests {
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation2(&vec![accounts.get(0).unwrap().clone()], false, true, &None);
+        let estimates = calculation2(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            false,
+            &Some(150),
+            &tracker,
+        );
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
@@ -993,9 +536,36 @@ mod tests {
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
+        // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
+        let estimates = calculation2(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            true,
+            &None,
+            &tracker,
+        );
+        // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
+        let expected_min_fee = 0.0;
+        let expected_low_fee = 25.0;
+        let expected_medium_fee = 50.0;
+        let expected_high_fee = 75.0;
+        let expected_very_high_fee = 95.0;
+        let expected_max_fee = 100.0;
+        assert_eq!(estimates.min, expected_min_fee);
+        assert_eq!(estimates.low, expected_low_fee);
+        assert_eq!(estimates.medium, expected_medium_fee);
+        assert_eq!(estimates.high, expected_high_fee);
+        assert_eq!(estimates.very_high, expected_very_high_fee);
+        assert_eq!(estimates.unsafe_max, expected_max_fee);
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation2(&vec![accounts.get(0).unwrap().clone()], false, true, &Some(150));
+        let estimates = calculation2(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            true,
+            &Some(150),
+            &tracker,
+        );
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
@@ -1035,7 +605,13 @@ mod tests {
         }
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation1(&vec![accounts.get(0).unwrap().clone()], false, false,  &None);
+        let estimates = calculation1(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            false,
+            &None,
+            &tracker,
+        );
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
@@ -1049,9 +625,14 @@ mod tests {
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
-
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation1(&vec![accounts.get(0).unwrap().clone()], false, false,  &Some(150));
+        let estimates = calculation1(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            false,
+            &Some(150),
+            &tracker,
+        );
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
@@ -1065,9 +646,14 @@ mod tests {
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
-
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation1(&vec![accounts.get(0).unwrap().clone()], false, true,  &None);
+        let estimates = calculation1(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            true,
+            &None,
+            &tracker,
+        );
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
@@ -1081,9 +667,14 @@ mod tests {
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
-
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation1(&vec![accounts.get(0).unwrap().clone()], false, true,  &Some(150));
+        let estimates = calculation1(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            true,
+            &Some(150),
+            &tracker,
+        );
         let expected_min_fee = 0.0;
         let expected_low_fee = 0.0;
         let expected_medium_fee = 25.5;
@@ -1121,7 +712,13 @@ mod tests {
         }
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation2(&vec![accounts.get(0).unwrap().clone()], false, false,  &None);
+        let estimates = calculation2(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            false,
+            &None,
+            &tracker,
+        );
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
@@ -1134,10 +731,15 @@ mod tests {
         assert_eq!(estimates.high, expected_high_fee);
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
-
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation2(&vec![accounts.get(0).unwrap().clone()], false, false,  &Some(150));
+        let estimates = calculation2(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            false,
+            &Some(150),
+            &tracker,
+        );
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
@@ -1150,10 +752,15 @@ mod tests {
         assert_eq!(estimates.high, expected_high_fee);
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
-
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation2(&vec![accounts.get(0).unwrap().clone()], false, true,  &None);
+        let estimates = calculation2(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            true,
+            &None,
+            &tracker,
+        );
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
@@ -1166,10 +773,15 @@ mod tests {
         assert_eq!(estimates.high, expected_high_fee);
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
-
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation2(&vec![accounts.get(0).unwrap().clone()], false, true,  &Some(150));
+        let estimates = calculation2(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            true,
+            &Some(150),
+            &tracker,
+        );
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
@@ -1182,7 +794,6 @@ mod tests {
         assert_eq!(estimates.high, expected_high_fee);
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
-
     }
 
     #[tokio::test]
@@ -1208,15 +819,25 @@ mod tests {
         // Simulate adding the fixed fees as both account-specific and transaction fees
         for (i, fee) in fees.clone().into_iter().enumerate() {
             if 0 == i.rem_euclid(10usize) {
-                tracker.push_priority_fee_for_txn(i as Slot, vec![account_unrelated], fee as u64, false);
-            }
-            else {
+                tracker.push_priority_fee_for_txn(
+                    i as Slot,
+                    vec![account_unrelated],
+                    fee as u64,
+                    false,
+                );
+            } else {
                 tracker.push_priority_fee_for_txn(i as Slot, accounts.clone(), fee as u64, false);
             }
         }
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation1(&vec![accounts.get(0).unwrap().clone()], false, false, &None);
+        let estimates = calculation1(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            false,
+            &None,
+            &tracker,
+        );
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
         let expected_high_fee = 75.0;
@@ -1227,7 +848,13 @@ mod tests {
         assert_ne!(estimates.very_high, expected_very_high_fee);
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation1(&vec![accounts.get(0).unwrap().clone()], false, false, &Some(150));
+        let estimates = calculation1(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            false,
+            &Some(150),
+            &tracker,
+        );
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
         let expected_high_fee = 75.0;
@@ -1238,7 +865,13 @@ mod tests {
         assert_ne!(estimates.very_high, expected_very_high_fee);
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation1(&vec![accounts.get(0).unwrap().clone()], false, true, &None);
+        let estimates = calculation1(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            true,
+            &None,
+            &tracker,
+        );
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
         let expected_high_fee = 75.0;
@@ -1249,7 +882,13 @@ mod tests {
         assert_ne!(estimates.very_high, expected_very_high_fee);
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation1(&vec![accounts.get(0).unwrap().clone()], false, true, &Some(150));
+        let estimates = calculation1(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            true,
+            &Some(150),
+            &tracker,
+        );
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
         let expected_high_fee = 75.0;
@@ -1259,7 +898,6 @@ mod tests {
         assert_ne!(estimates.high, expected_high_fee);
         assert_ne!(estimates.very_high, expected_very_high_fee);
     }
-
 
     #[tokio::test]
     async fn test_with_many_slots_broken_v2() {
@@ -1286,15 +924,25 @@ mod tests {
         // Simulate adding the fixed fees as both account-specific and transaction fees
         for (i, fee) in fees.clone().into_iter().enumerate() {
             if 0 == i.rem_euclid(10usize) {
-                tracker.push_priority_fee_for_txn(i as Slot, vec![account_unrelated], fee as u64, false);
-            }
-            else {
+                tracker.push_priority_fee_for_txn(
+                    i as Slot,
+                    vec![account_unrelated],
+                    fee as u64,
+                    false,
+                );
+            } else {
                 tracker.push_priority_fee_for_txn(i as Slot, accounts.clone(), fee as u64, false);
             }
         }
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation2(&vec![accounts.get(0).unwrap().clone()], false, false, &None);
+        let estimates = calculation2(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            false,
+            &None,
+            &tracker,
+        );
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
         let expected_high_fee = 75.0;
@@ -1305,7 +953,13 @@ mod tests {
         assert_ne!(estimates.very_high, expected_very_high_fee);
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation2(&vec![accounts.get(0).unwrap().clone()], false, false, &Some(150));
+        let estimates = calculation2(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            false,
+            &Some(150),
+            &tracker,
+        );
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
         let expected_high_fee = 75.0;
@@ -1316,7 +970,13 @@ mod tests {
         assert_ne!(estimates.very_high, expected_very_high_fee);
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation2(&vec![accounts.get(0).unwrap().clone()], false, true, &None);
+        let estimates = calculation2(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            true,
+            &None,
+            &tracker,
+        );
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
         let expected_high_fee = 75.0;
@@ -1327,7 +987,13 @@ mod tests {
         assert_ne!(estimates.very_high, expected_very_high_fee);
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
-        let estimates = tracker.calculation2(&vec![accounts.get(0).unwrap().clone()], false, true, &Some(150));
+        let estimates = calculation2(
+            &vec![accounts.get(0).unwrap().clone()],
+            false,
+            true,
+            &Some(150),
+            &tracker,
+        );
         let expected_low_fee = 25.0;
         let expected_medium_fee = 50.0;
         let expected_high_fee = 75.0;
@@ -1336,7 +1002,6 @@ mod tests {
         assert_ne!(estimates.medium, expected_medium_fee);
         assert_ne!(estimates.high, expected_high_fee);
         assert_ne!(estimates.very_high, expected_very_high_fee);
-
     }
 
     #[tokio::test]
@@ -1364,7 +1029,7 @@ mod tests {
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
         let v = vec![account_1];
-        let estimates = tracker.calculation1(&v, false, false, &None);
+        let estimates = calculation1(&v, false, false, &None, &tracker);
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
@@ -1379,8 +1044,7 @@ mod tests {
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
-
-        let estimates = tracker.calculation1(&v, false, false, &Some(150));
+        let estimates = calculation1(&v, false, false, &Some(150), &tracker);
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
@@ -1395,8 +1059,7 @@ mod tests {
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
-
-        let estimates = tracker.calculation1(&v, false, true, &None);
+        let estimates = calculation1(&v, false, true, &None, &tracker);
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
@@ -1411,8 +1074,7 @@ mod tests {
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
-
-        let estimates = tracker.calculation1(&v, false, true, &Some(150));
+        let estimates = calculation1(&v, false, true, &Some(150), &tracker);
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 0.0;
@@ -1453,7 +1115,7 @@ mod tests {
 
         // Now test the fee estimates for a known priority level, let's say medium (50th percentile)
         let v = vec![account_1];
-        let estimates = tracker.calculation2(&v, false, false, &None);
+        let estimates = calculation2(&v, false, false, &None, &tracker);
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
@@ -1468,8 +1130,7 @@ mod tests {
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
-
-        let estimates = tracker.calculation2(&v, false, false, &Some(150));
+        let estimates = calculation2(&v, false, false, &Some(150), &tracker);
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
@@ -1484,8 +1145,7 @@ mod tests {
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
-
-        let estimates = tracker.calculation2(&v, false, true, &None);
+        let estimates = calculation2(&v, false, true, &None, &tracker);
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
@@ -1500,8 +1160,7 @@ mod tests {
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
 
-
-        let estimates = tracker.calculation2(&v, false, true, &Some(150));
+        let estimates = calculation2(&v, false, true, &Some(150), &tracker);
         // Since the fixed fees are evenly distributed, the 50th percentile should be the middle value
         let expected_min_fee = 0.0;
         let expected_low_fee = 25.0;
@@ -1515,7 +1174,6 @@ mod tests {
         assert_eq!(estimates.high, expected_high_fee);
         assert_eq!(estimates.very_high, expected_very_high_fee);
         assert_eq!(estimates.unsafe_max, expected_max_fee);
-
     }
 
     #[test]
